cdk/full-chatbot/pyproject.toml
[tool.poetry]
name = "yuanta-chatbot-infrastructure"
version = "0.1.0"
description = "Full AI Chatbot Infrastructure using CDK Shared Constructs"
authors = ["Yuanta Team <team@yuanta.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "^2.90.0"
constructs = "^10.0.0"
boto3 = "^1.38.0"
python-dotenv = "^1.0.0"
cdk-shared-constructs = {path = "../../cdk_shared_constructs", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 